requirements.txt
Django==5.2.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
PyJWT==2.9.0
drf-yasg==1.21.7
Pillow==10.4.0
python-decouple==3.8